Minor bug fixes for server counts and empty/null column names

+++ src/main/java/org/mcstats/model/Server.java
@@ -129,7 +129,7 @@ public class Server implements Savable {
      * @return
      */
     public boolean recentlySentData() {
-        return lastSentData > (((int) System.currentTimeMillis() / 1000) - 1800);
+        return lastSentData > ((System.currentTimeMillis() / 1000) - 1800);
     }
 
     @Override
